updated migrations for price float to int

+++ app/database/migrations/2015_01_15_051642_create_restaurants_table.php
@@ -19,8 +19,8 @@ class CreateRestaurantsTable extends Migration {
 			$table->text('description', 500);
 			$table->integer('area_id');
 			$table->boolean('pure_veg');
-			$table->float('min_order');
-			$table->float('delivery_fee');
+			$table->integer('min_order');
+			$table->integer('delivery_fee');
 			$table->integer('delivery_time')->unsigned();
 			$table->boolean('online_payment');
 			$table->boolean('cash_on_delivery');
